requirements.txt
requests
beautifulsoup4
openai
python-dotenv
Pillow
gtts
pygame
easygui
SpeechRecognition
pydub
